Do not remove the event listeners of the sockets (#16)
* Do not remove the event listeners of the sockets

Sockets are not reused so the listeners should be GC'ed when the sockets
are closed. Furthermore there are cases where the `'error'` listeners
are prematurely removed leading to unhandled errors and unexpected
crashes.

Refs: https://github.com/TooTallNate/node-https-proxy-agent/pull/79

* Partially revert 58ed8f061e8f0ae05cde3ecdcebc9e58502214b4

Resuming the client socket is no longer needed as it is destroyed when
the target socket emits `'close'`.